create unified polling using checkForUpdates()
remove checkForPendingTransactionsUpdate as it was doing similar things but just unconfirmations, checkForUpdates had a bug which was not checking for unconfirms.. now it is so we merge in one.

Also fix setting active account twice in set active account model. FIxing eip1559 checks in utxo model if not bitcoin based. Fix setting account in setAccount, which was breaking sending funds when switching accounts.