Bug: Loading State Not Reset On Success
The isChanging loading state is set to true at the start of the handleChangeAccount function but is only reset to false in the catch block. If the account change operation succeeds, window.close() is called without first resetting isChanging. This can leave the UI stuck in a perpetual loading state if window.close() fails or is delayed. The loading state should be reset before window.close() or in a finally block.